pyproject.toml
[project]
name = "marquee-flight-tracker"
version = "0.1.0"
description = "Track overhead aircraft and display on a scrolling marquee"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.25",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
web = ["flask>=3.0"]
dev = ["pytest>=7.0"]

[project.scripts]
flight-tracker = "marquee_flight_tracker.__main__:main"
